pyproject.toml
[tool.poetry]
name = "tmdb"
version = "0.1.0"
description = "Movies/Favorites BFF in Django + DRF"
package-mode = false


[tool.poetry.dependencies]
python = ">=3.12,<4" #
django = "^5.2.7"
djangorestframework = "^3.16.1"
django-environ = "^0.12.0"
django-cors-headers = "^4.9.0"
django-filter = "^25.2"
drf-spectacular = "^0.28.0"
requests = "^2.32.5"
psycopg2-binary = "^2.9.11"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
[dependency-groups]
dev = [
    "pre-commit (>=4.3.0,<5.0.0)",
    "black (>=25.9.0,<26.0.0)",
    "isort (>=7.0.0,<8.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "mypy (>=1.18.2,<2.0.0)",
    "pytest (>=8.4.2,<9.0.0)",
    "pytest-django (>=4.11.1,<5.0.0)",
    "model-bakery (>=1.20.5,<2.0.0)",
    "django-stubs (>=5.2.7,<6.0.0)",
    "mypy-extensions (>=1.1.0,<2.0.0)",
    "types-requests (>=2.32.4.20250913,<3.0.0.0)"
]
